% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CamelNames.R
\name{CamelNames}
\alias{CamelNames}
\title{Convert column names to camelCase}
\usage{
CamelNames(names)
}
\arguments{
\item{names}{list of names as from dataframe colnames().}
}
\value{
A list of names converted to camel case
}
\description{
Convert column names to camelCase by taking them all to lower case, and looking for word boundaries and makeing those upper case.
Then trim off the white space.
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
