% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{Volcano}
\alias{Volcano}
\title{Create a Volcano plot}
\usage{
Volcano(fit, target, title, numLabels = 50, write = FALSE,
  saveImage = FALSE, ext = "png", cutFC = 1.5, cutPvalue = 0.05,
  xlim = NULL, labelSource = "hgnc_symbol",
  repelFontFace = theme_get()$text$face, repelSize = 4,
  pointSize = 3)
}
\arguments{
\item{fit}{fit ebays based fit after processing with limma}

\item{target}{string Name of the constrast target to plot}

\item{title}{string Title for the plot}

\item{numLabels}{integer Number of gene labels to show on the plot}

\item{write}{boolean Write DEG list (using contrast target string)}

\item{saveImage}{boolean Write out an image}

\item{ext}{string Image type/extension}

\item{labelSource}{string Name of the column for gene names}

\item{repelSize}{integer Size of labels}

\item{pointSize}{integet Size of points}
}
\value{
dataFrame Data Frame of differentially expressed gene information, in the form of topTable().
}
\description{
Make a volcano plot, using cuts, ggrepel, optionally write out as image, optionally write out DEG list and all gene list to .csv
}
\examples{
Volcano(fit, "SPvsDP", "Single Positive vs Double Positive")


}
