% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{Volcano}
\alias{Volcano}
\title{Create a Volcano plot}
\usage{
Volcano(fit, target, title, numLabels = 50, write = FALSE,
  writeTopTable = FALSE, saveImage = FALSE, ext = "png",
  extImage = "png", cutFC = 1.5, cutPvalue = 0.05,
  labelList = NULL, labelSize = 4, labelSource = "hgnc_symbol",
  colors = c("red", "blue", "grey"),
  repelFontFace = theme_get()$text$face, repelSize = 4,
  pointSize = 3, ...)
}
\arguments{
\item{fit}{A fitted linear model object (MArrayLM)}

\item{target}{Name of the contrast target to plot, or the corresponding column of the model matrix (string)}

\item{title}{Title for the plot (string)}

\item{numLabels}{Number of labels to print (integer)}

\item{write}{deprecated in favor of \code{writeTopTable}}

\item{writeTopTable}{Write gene list to CSV file (see \code{target} string, boolean)}

\item{saveImage}{Write out an image (boolean)}

\item{ext}{deprecated in favor of \code{extImage}}

\item{extImage}{Image type/extension (string)}

\item{cutFC}{Fold Change cutoff for significance (floating point)}

\item{cutPvalue}{p-valut cutoff for significance (floating point)}

\item{labelList}{List of label names to render, only those (vector)}

\item{labelSize}{Size of text in repel for labels from labelList - default is the same as the repelSize (integer)}

\item{labelSource}{Name of the column for gene names (string)}

\item{colors}{vector of up/right point color, down/left point color, and not significant color (vector of three strings)}

\item{repelFontFace}{Name of fontface for labels (string)}

\item{repelSize}{Size of text repel labels (integer)}

\item{pointSize}{Size of geom_points (integer)}

\item{...}{Additional function calls from the ggplot2 package to be added to ggplot, such as \code{xlim=xlim(-5,5)}, \code{ylim=ylim(0,9)}}
}
\value{
A dataframe of differentially expressed gene information, in the form of topTable().
}
\description{
Make a volcano plot, using cuts, ggrepel, optionally write out as image, optionally write out DEG list and all gene list to .csv
}
\examples{
\donttest{
Volcano(fit, "SPvsDP", "Single Positive vs Double Positive")
}


}
\seealso{
Associated or similar limma functions \code{\link[limma]{topTable}},  \code{\link[limma]{volcanoplot}}
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
