% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolcanoPlot.R
\name{Volcano}
\alias{Volcano}
\title{Create a Volcano plot}
\usage{
Volcano(fit, target, title, numLabels = 50, write,
  writeTopTable = FALSE, saveImage = FALSE, ext, extImage = "png",
  cutFC = 1.5, cutPvalue = 0.05, xlim = NULL, labelList = NULL,
  labelSource = "hgnc_symbol", downColor = "blue", nonColor = "grey",
  upColor = "red", repelFontFace = theme_get()$text$face,
  repelSize = 4, pointSize = 3)
}
\arguments{
\item{fit}{An MArrayLM fitted linear model object}

\item{target}{Name of the constrast target to plot (string)}

\item{title}{Title for the plot (string)}

\item{numLabels}{Number of gene labels to show on the plot (integer)}

\item{write}{deprecated in favor of \code{witeTopTable}}

\item{writeTopTable}{Write DEG list (using contrast target string, boolean)}

\item{saveImage}{Write out an image (boolean)}

\item{ext}{deprecated in favor of \code{extImage}}

\item{extImage}{Image type/extension (string)}

\item{cutFC}{Fold Change cutoff for significance (floating point)}

\item{cutPvalue}{p-valut cutoff for significance (floating point)}

\item{xlim}{Limits on x-axis (e.g. xlim(-5, 5))}

\item{labelList}{List of label names to render, only those (list)}

\item{labelSource}{Name of the column for gene names (string)}

\item{downColor}{Name of down/left point color (string)}

\item{nonColor}{Name of non-sigfificant color (string)}

\item{upColor}{name of up/right point color (string)}

\item{repelFontFace}{Name of fontface for labels (string)}

\item{repelSize}{Size of labels (integer)}

\item{pointSize}{Size of points (integer)}
}
\value{
A dataframe of differentially expressed gene information, in the form of topTable().
}
\description{
Make a volcano plot, using cuts, ggrepel, optionally write out as image, optionally write out DEG list and all gene list to .csv
}
\examples{
\donttest{
Volcano(fit, "SPvsDP", "Single Positive vs Double Positive")
}


}
\seealso{
Associated or similar limma functions \code{\link[limma]{topTable}},  \code{\link[limma]{volcanoplot}}
}
\author{
Scott R Presnell, \email{SPresnell@benaroyaresearch.org}
}
